+++ src/script/mod.rs
@@ -1,1 +1,2 @@
-mod elements;
+mod parser;
+mod script;
